# src/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OHC账票生成服务",
    "description": "OHC账票生成API服务",
    "version": "1.0.0"
  },
  "paths": {
...
    "/generate": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成文档",
        "description": "生成账票文档（通用接口）",
        "operationId": "generate_document_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/dhf-index": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成DHF INDEX",
        "description": "生成制作文档・图纸一览",
        "operationId": "generate_dhf_index_generate_dhf_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DHFIndexParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/ptf-index": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成PTF INDEX",
        "description": "生成PTF INDEX",
        "operationId": "generate_ptf_index_generate_ptf_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PTFIndexParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/es-individual-test-spec": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成ES个别试验要项书",
        "description": "生成ES个别试验要项书",
        "operationId": "generate_es_individual_test_spec_generate_es_individual_test_spec_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ESIndividualTestSpecParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/es-individual-test-result": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成ES个别试验结果书",
        "description": "生成ES个别试验结果书",
        "operationId": "generate_es_individual_test_result_generate_es_individual_test_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ESIndividualTestResultParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/pp-individual-test-result": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成PP个别试验结果书",
        "description": "生成PP个别试验结果书",
        "operationId": "generate_pp_individual_test_result_generate_pp_individual_test_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PPIndividualTestResultParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/es-verification-plan": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成ES验证计划书",
        "description": "生成ES验证计划书",
        "operationId": "generate_es_verification_plan_generate_es_verification_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ESVerificationPlanParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/es-verification-result": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成ES验证结果书",
        "description": "生成ES验证结果书",
        "operationId": "generate_es_verification_result_generate_es_verification_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ESVerificationResultParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/pp-verification-plan": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成PP验证计划书",
        "description": "生成PP验证计划书",
        "operationId": "generate_pp_verification_plan_generate_pp_verification_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PPVerificationPlanParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/pp-verification-result": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成PP验证结果书",
        "description": "生成PP验证结果书",
        "operationId": "generate_pp_verification_result_generate_pp_verification_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PPVerificationResultParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/basic-specification": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成基本规格书",
        "description": "生成基本规格书",
        "operationId": "generate_basic_specification_generate_basic_specification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicSpecificationParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/pp-individual-test-spec": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成PP个别试验要项书",
        "description": "生成PP个别试验要项书",
        "operationId": "generate_pp_individual_test_spec_generate_pp_individual_test_spec_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PPIndividualTestSpecParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/follow-up-dr-minutes": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成跟进DR会议记录",
        "description": "生成跟进DR会议记录",
        "operationId": "generate_follow_up_dr_minutes_generate_follow_up_dr_minutes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpDRMinutesParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/labeling-specification": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成标签规格书",
        "description": "生成标签规格书",
        "operationId": "generate_labeling_specification_generate_labeling_specification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelingSpecificationParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/product-environment-assessment": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成产品环境评估要项书/结果书",
        "description": "生成产品环境评估要项书/结果书",
        "operationId": "generate_product_environment_assessment_generate_product_environment_assessment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductEnvironmentAssessmentParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/existing-product-comparison": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成与现有产品对比表",
        "description": "生成与现有产品对比表",
        "operationId": "generate_existing_product_comparison_generate_existing_product_comparison_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExistingProductComparisonParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/packaging-design-specification": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成包装设计仕样书",
        "description": "生成包装设计仕样书",
        "operationId": "generate_packaging_design_specification_generate_packaging_design_specification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackagingDesignSpecificationParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/user-manual-specification": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成使用说明书仕样书",
        "description": "生成使用说明书仕样书",
        "operationId": "generate_user_manual_specification_generate_user_manual_specification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserManualSpecificationParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/project-plan": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "生成项目计划书",
        "description": "生成项目计划书",
        "operationId": "generate_project_plan_generate_project_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPlanParameters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateDocumentRequest": {
        "properties": {
          "template_name": {
            "type": "string",
            "title": "Template Name",
            "description": "模板名称"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "模板参数"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "语言代码 (zh/ja/en)，如果不指定则使用默认模板"
          }
        },
        "type": "object",
        "required": [
          "template_name",
          "parameters"
        ],
        "title": "GenerateDocumentRequest",
        "description": "生成文档请求模型",
        "example": {
          "language": "zh",
          "parameters": {
            "author": "张三",
            "date": "2025-01-22",
            "department": "研发部",
            "document_type": "设计文档",
            "project_name": "OHC项目",
            "version": "1.0"
          },
          "template_name": "DHF_INDEX"
        }
      },
      "GenerateDocumentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "是否成功"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name",
            "description": "生成的文件名"
          },
          "file_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Url",
            "description": "文件下载链接（远程存储时）"
          },
          "file_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Content",
            "description": "文件内容（Base64编码，本地存储时）"
          },
          "storage_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Type",
            "description": "存储类型"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id",
            "description": "项目编号"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "版本号"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "GenerateDocumentResponse",
        "description": "生成文档响应模型",
        "example": {
          "file_content": "UEsDBBQAAAAIAA...",
          "file_name": "OHC项目_v1.0_20250122_DHF_INDEX.xlsx",
          "message": "文档生成成功",
          "project_id": "OHC项目",
          "storage_type": "local",
          "success": true,
          "version": "1.0"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}